cleaning up and moving around functions
